+++ packages/iov-bcp/src/atomicswaptypes.ts
@@ -75,19 +75,18 @@ export interface AtomicSwapSenderQuery {
 }
 
 export interface AtomicSwapIdQuery {
-  readonly swapid: SwapId;
+  readonly id: SwapId;
 }
 
-// on some chains, swapid may equal hashlock, but often these may differ
-export interface AtomicSwapHashlockQuery {
-  readonly hashlock: Uint8Array;
+export interface AtomicSwapHashQuery {
+  readonly hash: Hash;
 }
 
 export type AtomicSwapQuery =
   | AtomicSwapRecipientQuery
   | AtomicSwapSenderQuery
   | AtomicSwapIdQuery
-  | AtomicSwapHashlockQuery;
+  | AtomicSwapHashQuery;
 
 /** a type guard to use in the swap-based queries  */
 export function isAtomicSwapRecipientQuery(query: AtomicSwapQuery): query is AtomicSwapRecipientQuery {
@@ -101,12 +100,12 @@ export function isAtomicSwapSenderQuery(query: AtomicSwapQuery): query is Atomic
 
 /** a type guard to use in the swap-based queries  */
 export function isAtomicSwapIdQuery(query: AtomicSwapQuery): query is AtomicSwapIdQuery {
-  return (query as AtomicSwapIdQuery).swapid !== undefined;
+  return (query as AtomicSwapIdQuery).id !== undefined;
 }
 
 /** a type guard to use in the swap-based queries  */
-export function isAtomicSwapHashlockQuery(query: AtomicSwapQuery): query is AtomicSwapHashlockQuery {
-  return (query as AtomicSwapHashlockQuery).hashlock !== undefined;
+export function isAtomicSwapHashQuery(query: AtomicSwapQuery): query is AtomicSwapHashQuery {
+  return (query as AtomicSwapHashQuery).hash !== undefined;
 }
 
 /**