+++ packages/iov-bns/src/bnsconnection.spec.ts
@@ -1558,13 +1558,13 @@ describe("BnsConnection", () => {
 
     // ----  prepare queries
     const querySwapId: AtomicSwapQuery = {
-      swapid: {
+      id: {
         data: txResult,
       },
     };
     const querySwapSender: AtomicSwapQuery = { sender: faucetAddr };
     const querySwapRecipient: AtomicSwapQuery = { recipient: recipientAddr };
-    const querySwapHash: AtomicSwapQuery = { hashlock: swapOfferHash };
+    const querySwapHash: AtomicSwapQuery = { hash: swapOfferHash };
 
     // ----- connection.searchTx() -----
     // we should be able to find the transaction through quite a number of tag queries