+++ packages/iov-bns/src/tags.ts
@@ -13,14 +13,14 @@ export function bnsSwapQueryTag(query: AtomicSwapQuery, set = true): QueryTag {
   let binKey: Uint8Array;
   const bucket = "esc";
   if (isAtomicSwapIdQuery(query)) {
-    binKey = Uint8Array.from([...bucketKey(bucket), ...query.swapid.data]);
+    binKey = Uint8Array.from([...bucketKey(bucket), ...query.id.data]);
   } else if (isAtomicSwapSenderQuery(query)) {
     binKey = Uint8Array.from([...indexKey(bucket, "sender"), ...decodeBnsAddress(query.sender).data]);
   } else if (isAtomicSwapRecipientQuery(query)) {
     binKey = Uint8Array.from([...indexKey(bucket, "recipient"), ...decodeBnsAddress(query.recipient).data]);
   } else {
     // if (isQueryBySwapHash(query))
-    binKey = Uint8Array.from([...indexKey(bucket, "arbiter"), ...hashIdentifier(query.hashlock)]);
+    binKey = Uint8Array.from([...indexKey(bucket, "arbiter"), ...hashIdentifier(query.hash)]);
   }
 
   return {