+++ packages/iov-ethereum/src/ethereumconnection.ts
@@ -22,7 +22,7 @@ import {
   FailedTransaction,
   Fee,
   Hash,
-  isAtomicSwapHashlockQuery,
+  isAtomicSwapHashQuery,
   isAtomicSwapIdQuery,
   isAtomicSwapRecipientQuery,
   isAtomicSwapSenderQuery,
@@ -781,7 +781,7 @@ export class EthereumConnection implements AtomicSwapConnection {
     maxHeight: number = Number.MAX_SAFE_INTEGER,
   ): Promise<ReadonlyArray<AtomicSwap>> {
     if (isAtomicSwapIdQuery(query)) {
-      const data = Uint8Array.from([...Abi.calculateMethodId("get(bytes32)"), ...query.swapid.data]);
+      const data = Uint8Array.from([...Abi.calculateMethodId("get(bytes32)"), ...query.id.data]);
 
       const params = [
         {
@@ -820,7 +820,7 @@ export class EthereumConnection implements AtomicSwapConnection {
 
       const resultArray = Encoding.fromHex(normalizeHex(swapsResponse.result));
       const swapData: SwapData = {
-        id: query.swapid,
+        id: query.id,
         sender: toChecksummedAddress(Abi.decodeAddress(resultArray.slice(senderBegin, senderEnd))),
         recipient: toChecksummedAddress(Abi.decodeAddress(resultArray.slice(recipientBegin, recipientEnd))),
         hash: resultArray.slice(hashBegin, hashEnd) as Hash,
@@ -862,7 +862,7 @@ export class EthereumConnection implements AtomicSwapConnection {
     } else if (
       isAtomicSwapRecipientQuery(query) ||
       isAtomicSwapSenderQuery(query) ||
-      isAtomicSwapHashlockQuery(query)
+      isAtomicSwapHashQuery(query)
     ) {
       const params = [
         {
@@ -975,8 +975,8 @@ export class EthereumConnection implements AtomicSwapConnection {
               return swap.data.recipient === query.recipient;
             } else if (isAtomicSwapSenderQuery(query)) {
               return swap.data.sender === query.sender;
-            } else if (isAtomicSwapHashlockQuery(query)) {
-              return Encoding.toHex(swap.data.hash) === Encoding.toHex(query.hashlock);
+            } else if (isAtomicSwapHashQuery(query)) {
+              return Encoding.toHex(swap.data.hash) === Encoding.toHex(query.hash);
             }
             throw new Error("unsupported query type");
           },