+++ CHANGELOG.md
@@ -2,14 +2,21 @@
 
 ## 0.15.0
 
+- @iov/ethereum: Add `createEtherSwapId` and `createErc20SwapId` static methods.
+
+Breaking changes
+
 - @iov/bcp: Remove unused type `SwapIdString`.
 - @iov/bcp: Remove `AtomicSwapHelpers.createId`, which is now the responsibility
   of individual blockchain connections.
 - @iov/bcp: Add `SwapId` type, and use where appropriate. This is an interface
   containing `SwapIdBytes` and can be extended on a per-chain basis.
+- @iov/bcp: Rename `AtomicSwapIdQuery.swapid` to `AtomicSwapIdQuery.id`.
+- @iov/bcp: Rename `AtomicSwapHashlockQuery.hashlock` to
+  `AtomicSwapHashQuery.hash` and change to type `Hash`; rename
+  `isAtomicSwapHashlockQuery` to `isAtomicSwapHashQuery`.
 - @iov/ethereum: Switch to new `SwapId` type instead of `SwapIdBytes` where
   appropriate.
-- @iov/ethereum: Add `createEtherSwapId` and `createErc20SwapId` static methods.
 - @iov/bns: Switch to new `SwapId` type instead of `SwapIdBytes` where
   appropriate.